[REFACTOR] extract redirect for authenticated users into composable
